feat: #467 - Move some ClassDeclaration functionality into ClassLikeDeclarationBase.
This will be useful for supporting ClassExpression.
