+++ scripts/generation/main.ts
@@ -1,4 +1,5 @@
-﻿import { createGetStructureFunctions } from "./createGetStructureFunctions";
+﻿import { ts } from "ts-simple-ast";
+import { createGetStructureFunctions } from "./createGetStructureFunctions";
 import { createTypeGuardsUtility } from "./createTypeGuardsUtility";
 import { createCodeBlockWriterFile } from "./createCodeBlockWriterFile";
 import { createKindToNodeMappings } from "./createKindToNodeMappings";
@@ -12,6 +13,8 @@ const originalArgs = [...args];
 const factory = new InspectorFactory();
 const inspector = factory.getTsSimpleAstInspector();
 const tsInspector = factory.getTsInspector();
+
+console.log(`TypeScript version: ${ts.version}`);
 
 if (checkHasArg("create-get-structure-functions")) {
     console.log("Creating get structure functions...");
