+++ src/compiler/ast/class/base/index.ts
@@ -1,1 +1,2 @@
 export * from "./AbstractableNode";
+export * from "./ClassLikeDeclarationBase";
